IMS_Backend/requirements.txt
Django==4.2.29
djangorestframework==3.17.1
django-filter==24.3
djangorestframework-simplejwt==5.5.1
psycopg[binary]==3.2.9
django-cors-headers==4.9.0
gunicorn==23.0.0
whitenoise==6.8.2
openpyxl==3.1.5
reportlab==4.2.5
